refactor: dns/server package is now dns package
