+++ client.go
@@ -3,7 +3,7 @@ package gowindows
 
 import (
 	"github.com/d-strobel/gowindows/connection"
-	"github.com/d-strobel/gowindows/windows/dns/server"
+	"github.com/d-strobel/gowindows/windows/dns"
 	"github.com/d-strobel/gowindows/windows/local/accounts"
 )
 
@@ -15,7 +15,7 @@ import (
 type Client struct {
 	Connection    connection.Connection
 	LocalAccounts *accounts.Client
-	DnsServer     *server.Client
+	Dns           *dns.Client
 }
 
 // NewClient returns a new instance of the Client object, initialized with the provided configuration.
@@ -28,7 +28,7 @@ func NewClient(conn connection.Connection) *Client {
 
 	// Build the client with the subpackages.
 	c.LocalAccounts = accounts.NewClient(c.Connection)
-	c.DnsServer = server.NewClient(c.Connection)
+	c.Dns = dns.NewClient(c.Connection)
 
 	return c
 }
