+++ windows/dns/record_aaaa_test.go
@@ -1,4 +1,4 @@
-package server
+package dns
 
 import (
 	"context"
