+++ windows/dns/record_cname_test.go
@@ -1,4 +1,4 @@
-package server
+package dns
 
 import (
 	"context"
